refactor: 필요 없는 파일 삭제 및 기본 컨벤션 적용
