+++ src/main/java/lotto/input/LottoPrice.java
@@ -1,6 +1,7 @@
 package lotto.input;
 
 public class LottoPrice {
+
     public static final int LOTTO_PRICE = 1000;
     private final int lottoPrice;
