--- src/main/java/lotto/lottos/NumberGenerator/NumbersGenerator.java
@@ -1,8 +1,0 @@
-package lotto.lottos.NumberGenerator;
-
-import java.util.List;
-
-public interface NumbersGenerator {
-
-    List<Integer> generate();
-}
